refactor(agent): remove legacy skill prompts and inert limits
Retire the unused eager/index renderers and rollout flag now that Factory owns native prompt preparation. Preserve spec conversion tests and remove no-op warning thresholds and unenforced bundle defaults.

Refs: zaq-8q7.1, zaq-8q7.2
