+++ test/zaq/agent/skills_test.exs
@@ -289,59 +289,4 @@ defmodule Zaq.Agent.SkillsTest do
       assert Skills.provisioned_mcp_endpoint_ids(agent, [skill]) == []
     end
   end
-
-  describe "effective_system_prompt/2 and render_prompt_block/1" do
-    test "returns the bare job when there are no skills" do
-      agent = %ConfiguredAgent{job: "You are helpful."}
-      assert Skills.effective_system_prompt(agent, []) == "You are helpful."
-    end
-
-    test "appends the rendered skills block to the job" do
-      agent = %ConfiguredAgent{job: "You are helpful."}
-
-      skill = %Skill{
-        name: "calculator",
-        description: "Precise arithmetic",
-        body: "Always use tools for math."
-      }
-
-      prompt = Skills.effective_system_prompt(agent, [skill])
-
-      assert prompt =~ "You are helpful.\n\nYou have access to the following skills:"
-      assert prompt =~ "## calculator"
-      assert prompt =~ "Precise arithmetic"
-      assert prompt =~ "Always use tools for math."
-    end
-
-    test "returns just the skills block when the job is empty" do
-      skill = %Skill{name: "solo", body: "Body only."}
-
-      for agent <- [%ConfiguredAgent{job: nil}, %ConfiguredAgent{job: ""}] do
-        prompt = Skills.effective_system_prompt(agent, [skill])
-        assert String.starts_with?(prompt, "You have access to the following skills:")
-        assert prompt =~ "## solo"
-      end
-    end
-
-    test "render_prompt_block/1 returns nil for an empty list" do
-      assert Skills.render_prompt_block([]) == nil
-    end
-
-    test "render_prompt_block/1 skips blank descriptions" do
-      block = Skills.render_prompt_block([%Skill{name: "terse", description: nil, body: "B."}])
-      assert block == "You have access to the following skills:\n\n## terse\n\nB."
-    end
-
-    test "render_prompt_block/1 renders multiple skills in order" do
-      block =
-        Skills.render_prompt_block([
-          %Skill{name: "first", body: "One."},
-          %Skill{name: "second", body: "Two."}
-        ])
-
-      {first_idx, _} = :binary.match(block, "## first")
-      {second_idx, _} = :binary.match(block, "## second")
-      assert first_idx < second_idx
-    end
-  end
 end
